Implement change in draft 20 vs draft 17 #34

+++ dots_server/listen_db.go
@@ -17,6 +17,12 @@ import (
 	dots_config "github.com/nttdots/go-dots/dots_server/config"
 )
 
+type TableName string
+const (
+	MITIGATION_SCOPE      TableName = "mitigation_scope"
+	SESSION_CONFIGURATION TableName = "signal_session_configuration"
+)
+
 /*
  * Listen for notification from DB
  */
@@ -62,7 +68,7 @@ ILOOP:
 		case io.EOF:
 			break ILOOP
 		case nil:
-			log.Debugf("[MySQL-Notification]: Received mitigation status changed notification from DB for :", data)
+			log.Debugf("[MySQL-Notification]: Received data changed notification from DB trigger for: %+v", data)
 			if isTransportOver(data) {
 				break ILOOP
 			}
@@ -74,42 +80,52 @@ ILOOP:
 				log.Errorf("[MySQL-Notification]: Failed to encode json message to map data.")
 				return
 			}
-			id, iErr := strconv.ParseInt(mapData["id"].(string), 10, 64)
-			cid, cErr := strconv.Atoi(mapData["cid"].(string))
-			cuid := mapData["cuid"].(string)
-			mid, mErr := strconv.Atoi(mapData["mid"].(string))
-			status, sErr := strconv.Atoi(mapData["status"].(string))
-			if iErr != nil || mErr != nil || sErr != nil || cErr != nil {
-				log.Debugf("[MySQL-Notification]:Failed to parse string to integer")
-				return
-			}
-			uriPath := messages.MessageTypes[messages.MITIGATION_REQUEST].Path
-			query := uriPath + "/cuid=" + cuid + "/mid=" + strconv.Itoa(mid)
-
-			// Check duplicate mitigation when PUT a new mitigation before delete an expired mitigation
-			mids, err := models.GetMitigationIds(cid, cuid)
-			if err != nil {
-				log.Errorf("[MySQL-Notification]: Error: %+v", err)
-				return
-			}
-			dup := isDuplicateMitigation(mids, mid)
-			if dup && status == models.Terminated {
-				// Skip notify, just delete the expired mitigation
-				log.Debugf("[MySQL-Notification]: Skip Notification for this mitigation (mid=%+v, id=%+v) due to duplicate with another existing active mitigation", mid, id)
-			} else {
-				// Notify status changed to those clients who are observing this mitigation request
-				log.Debug("[MySQL-Notification]: Send notification if obsevers exists")
-				context.NotifyOnce(query)
-			}
-
-			// If mitigation status was changed to 6: (attack mitigation is now terminated), delete this mitigation after notifying
-			if status == models.Terminated {
-				log.Debugf("[MySQL-Notification]: Mitigation was terminated. Delete corresponding sub-resource and mitigation request.", models.Terminated)
-				controllers.DeleteMitigation(cid, cuid, mid, id)
-				// Keep resource when there is a duplication
-				if !dup {
-					context.DeleteResourceByQuery(query)
+			if mapData["table_trigger"].(string) == string(MITIGATION_SCOPE) {
+				id, iErr := strconv.ParseInt(mapData["id"].(string), 10, 64)
+				cid, cErr := strconv.Atoi(mapData["cid"].(string))
+				cuid := mapData["cuid"].(string)
+				mid, mErr := strconv.Atoi(mapData["mid"].(string))
+				status, sErr := strconv.Atoi(mapData["status"].(string))
+				if iErr != nil || mErr != nil || sErr != nil || cErr != nil {
+					log.Debugf("[MySQL-Notification]:Failed to parse string to integer")
+					return
 				}
+				uriPath := messages.MessageTypes[messages.MITIGATION_REQUEST].Path
+				query := uriPath + "/cuid=" + cuid + "/mid=" + strconv.Itoa(mid)
+
+				// Check duplicate mitigation when PUT a new mitigation before delete an expired mitigation
+				mids, err := models.GetMitigationIds(cid, cuid)
+				if err != nil {
+					log.Errorf("[MySQL-Notification]: Error: %+v", err)
+					return
+				}
+				dup := isDuplicateMitigation(mids, mid)
+				if dup && status == models.Terminated {
+					// Skip notify, just delete the expired mitigation
+					log.Debugf("[MySQL-Notification]: Skip Notification for this mitigation (mid=%+v, id=%+v) due to duplicate with another existing active mitigation", mid, id)
+				} else {
+					// Notify status changed to those clients who are observing this mitigation request
+					log.Debug("[MySQL-Notification]: Send notification if obsevers exists")
+					context.NotifyOnce(query)
+				}
+
+				// If mitigation status was changed to 6: (attack mitigation is now terminated), delete this mitigation after notifying
+				if status == models.Terminated {
+					log.Debugf("[MySQL-Notification]: Mitigation was terminated. Delete corresponding sub-resource and mitigation request.", models.Terminated)
+					controllers.DeleteMitigation(cid, cuid, mid, id)
+					// Keep resource when there is a duplication
+					if !dup {
+						context.DeleteResourceByQuery(query)
+					}
+				}
+			} else if mapData["table_trigger"].(string) == string(SESSION_CONFIGURATION) {
+
+				// Notify session configuration changed to those clients who are observing this mitigation request
+				log.Debug("[MySQL-Notification]: Send notification if obsevers exists")
+				cid := mapData["cid"].(string)
+				uriPath := messages.MessageTypes[messages.SESSION_CONFIGURATION].Path
+				query := uriPath + "/customerId=" + cid
+				context.NotifyOnce(query)
 			}
 		default:
 			log.Debugf("[MySQL-Notification]: Failed to receive data:%+v", err)
